Use u16 for remaster version

+++ 3gx/includes/title_info.h
@@ -3,4 +3,4 @@
 #include <3ds.h>
 
 u64 get_title_id();
-u64 get_remaster_version();
+u16 get_remaster_version();
