+++ reader_core/src/pnp/bindings.rs
@@ -7,7 +7,7 @@ extern "C" {
     pub fn host_get_game_title_id() -> u64;
     pub fn host_game_start_ms() -> u64;
     pub fn get_current_keys() -> u32;
-    pub fn get_remaster_version() -> u64;
+    pub fn get_remaster_version() -> u16;
     pub fn osGetTime() -> u64;
     pub fn get_trampoline_addr() -> u32;
     pub fn get_route_hook_addr() -> u32;
@@ -44,7 +44,7 @@ pub mod test_stubs {
         0
     }
     #[no_mangle]
-    pub extern "C" fn get_remaster_version() -> u64 {
+    pub extern "C" fn get_remaster_version() -> u16 {
         0
     }
     #[no_mangle]
